Move all the basic openscad objects into a new file
This almost halves the size of the main file.
Note that a circular import is necessary, since the base class references the
subclasses
